// src/types/auth.types.ts
export interface User {
  id: string;
  email: string;
  name: string;
  roles?: string[];
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface BackendLoginResponse {
  access_token: string;
}